[#83] Separate buildings that have floors and those who have not
