+++ components/map/building/buildingNoFloors/styles.js
@@ -13,14 +13,6 @@ const styles = StyleSheet.create({
     backgroundColor: 'white',
     position: 'absolute'
   },
-
-  buildingContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    position: 'absolute',
-    zIndex: 300
-  },
-
   quitButton: {
     left: 0,
     top: 100,
@@ -29,20 +21,6 @@ const styles = StyleSheet.create({
     height: 20,
     fontSize: 20
   },
-  switcher: {
-    display: 'flex',
-    flexDirection: 'row',
-    left: 0,
-    bottom: 100,
-  },
-  lvl: {
-    marginBottom: 175,
-    padding: 10,
-    backgroundColor: '#FFE8D2',
-    borderColor: '#9CD3D7',
-    borderWidth: 3,
-    borderStyle: 'solid'
-  }
 });
 
 export default styles;
